+++ src/components/Table/TableExpandRow.tsx
@@ -15,15 +15,20 @@ export interface TableExpandRowProps extends Omit<
      * This text is displayed as tooltip for the button that toggles the expanded/collapsed state.
      */
     togglerText: string;
+    /**
+     * Display this row with the styles from a zebra style-enabled table.
+     */
+    useZebraStyle?: boolean;
 };
 
 /**
  * Table row that is suffixed by a cell containing a button to expand/collapse this row.
  */
-function TableExpandRow ({
+export function TableExpandRow ({
     togglerText,
     isExpanded,
     isSelected,
+    useZebraStyle,
     onExpand,
     className,
     children,
@@ -43,6 +48,7 @@ function TableExpandRow ({
                 ` ${carbonClassPrefix}--parent-row` +
                 (isExpanded ? ` ${carbonClassPrefix}--expandable-row` : "") +
                 (isSelected ? ` ${carbonClassPrefix}--data-table--selected` : "") +
+                (useZebraStyle ? ` ${eccgui}-simpletable__row--zebra` : "") +
                 (!!className ? ` ${className}` : "")
             }
             data-parent-row={true}